+++ src/layers/ActivationBase.js
@@ -4,9 +4,9 @@
 	class ActivationBase {
 		constructor(size) {
 			this.nextLayer; //the connected layer
-			this.inData = new Float32Array(size);
-			this.outData = new Float32Array(size);
-			this.costs = new Float32Array(size); //costs for each neuron
+			this.inData = new Float64Array(size);
+			this.outData = new Float64Array(size);
+			this.costs = new Float64Array(size); //costs for each neuron
 			this.pl; //reference to previous layer
 		}
 
@@ -17,9 +17,9 @@
 			// try to connect to it
 			if (this.inData.length == 0) {
 				//if not already initialized
-				this.inData = new Float32Array(layer.outSize());
-				this.outData = new Float32Array(layer.outSize());
-				this.costs = new Float32Array(layer.outSize());
+				this.inData = new Float64Array(layer.outSize());
+				this.outData = new Float64Array(layer.outSize());
+				this.costs = new Float64Array(layer.outSize());
 			}
 			this.pl = layer;
 		}
