+++ src/layers/MinPoolLayer.js
@@ -28,13 +28,13 @@
 			this.filterWidth = filterWidth;
 			this.filterHeight = filterHeight;
 			this.stride = stride;
-			this.outData = new Float32Array(
+			this.outData = new Float64Array(
 				Math.ceil((inWidth - filterWidth + 1) / stride) * Math.ceil((inHeight - filterHeight + 1) / stride) * this.inDepth
 			);
-			this.inData = new Float32Array(inWidth * inHeight * inDepth);
-			this.costs = new Float32Array(inWidth * inHeight * inDepth);
-			this.minIndexes = new Float32Array(this.outData.length);
-			this.accessed = new Float32Array(this.costs.length).fill(1);
+			this.inData = new Float64Array(inWidth * inHeight * inDepth);
+			this.costs = new Float64Array(inWidth * inHeight * inDepth);
+			this.minIndexes = new Float64Array(this.outData.length);
+			this.accessed = new Float64Array(this.costs.length).fill(1);
 			if (this.filterWidth > inWidth || this.filterHeight > inHeight) {
 				throw "Min Pool layer error: Pooling size (width / height) cannot be bigger than the inputs corresponding (width/height)";
 			}
