+++ src/layers/RecEmitterLayer.js
@@ -4,9 +4,9 @@
 		constructor(id) {
 			this.id = id || 0;
 			this.nextLayer; //the connected layer
-			this.inData = new Float32Array(0); //the inData
-			this.outData = new Float32Array(0); //will be init when "connect" is called.
-			this.costs = new Float32Array(0); //costs for each neuron
+			this.inData = new Float64Array(0); //the inData
+			this.outData = new Float64Array(0); //will be init when "connect" is called.
+			this.costs = new Float64Array(0); //costs for each neuron
 			this.receiver; // a reference to the receiver layer so we can skip layers
 			//this will be set by the receiver  when the net is initialized
 			this.savedOutData;
@@ -21,12 +21,12 @@
 			if (layer.constructor.name == "RecReceiverLayer" && layer.id == this.id) {
 				throw "You can't put a RecReceiver right before its corressponding RecEmitter. (because it doesnt know the output size) (try adding a dummy layer inbetween and giving it a size)";
 			}
-			this.inData = new Float32Array(layer.outSize());
-			this.costs = new Float32Array(layer.outSize());
+			this.inData = new Float64Array(layer.outSize());
+			this.costs = new Float64Array(layer.outSize());
 			this.pl = layer;
 
-			this.outData = new Float32Array(layer.outSize());
-			this.savedOutData = new Float32Array(layer.outSize());
+			this.outData = new Float64Array(layer.outSize());
+			this.savedOutData = new Float64Array(layer.outSize());
 			this.savedOutData.fill(0);
 		}
